+++ referenced/indications.json
@@ -35,7 +35,7 @@
     {
         "id": 3,
         "document_id": 0,
-        "indication": "Verzenio is a kinase inhibitor indicated as monotherapy for the treatment of adult patients with HR-positive, HER2-negative advanced or metastatic breast cancer with disease progression following endocrine therapy and prior chemotherapy in the metastatic setting. ",
+        "indication": "Verzenio is a kinase inhibitor indicated as monotherapy for the treatment of adult patients with HR-positive, HER2-negative advanced or metastatic breast cancer with disease progression following endocrine therapy and prior chemotherapy in the metastatic setting.",
         "initial_approval_date": "2017-09-28",
         "initial_approval_url": "https://www.accessdata.fda.gov/drugsatfda_docs/label/2017/208716Orig1s000lbl.pdf",
         "description": "The U.S. Food and Drug Administration (FDA) granted approval to abemaciclib for the treatment of adult patients with hormone receptor (HR)-positive, human epidermal growth factor 2 (HER2)-negative advanced or metastatic breast cancer with disease progression following endocrine therapy and prior chemotherapy in the metastatic setting.",
@@ -679,7 +679,7 @@
     {
         "id": 61,
         "document_id": 27,
-        "indication": "SPRYCEL is a kinase inhibitor indicated for the treatment of pediatric patients 1 year of age and older with newly diagnosed Ph+ ALL in combination with chemotherapy. ",
+        "indication": "SPRYCEL is a kinase inhibitor indicated for the treatment of pediatric patients 1 year of age and older with newly diagnosed Ph+ ALL in combination with chemotherapy.",
         "initial_approval_date": "2018-12-21",
         "initial_approval_url": "https://www.accessdata.fda.gov/drugsatfda_docs/label/2018/021986s021lbl.pdf",
         "description": "The U.S. Food and Drug Administration granted approval to dasatinib in combination with chemotherapy for the treatment of pediatric patients 1 year of age and older with newly diagnosed Philadelphia chromosome-positive acute lymphoblastic leukemia (Ph+ ALL).",
@@ -1155,7 +1155,7 @@
         "indication": "TYKERB is a kinase inhibitor indicated in combination with capecitabine for the treatment of patients with advanced or metastatic breast cancer whose tumors overexpress human epidermal growth factor receptor 2 (HER2) and who have received prior therapy, including an anthracycline, a taxane, and trastuzumab. Limitations of Use: Patients should have disease progression on trastuzumab prior to initiation of treatment with TYKERB in combination with capecitabine. TYKERB in combination with an aromatase inhibitor has not been compared to a trastuzumab-containing chemotherapy regimen for the treatment of metastatic breast cancer.",
         "initial_approval_date": "2007-03-13",
         "initial_approval_url": "https://www.accessdata.fda.gov/drugsatfda_docs/label/2007/022059lbl.pdf",
-        "description": "The U.S. Food and Drug Administration granted approval to lapatinib in combination with capecitabine for the treatment of patients with advanced or metastatic breast cancer whose tumors overexpress human epidermal growth factor 2 (HER2) and who have received prior therapy, including an anthracycline, a taxane, a trastuzumab. ",
+        "description": "The U.S. Food and Drug Administration granted approval to lapatinib in combination with capecitabine for the treatment of patients with advanced or metastatic breast cancer whose tumors overexpress human epidermal growth factor 2 (HER2) and who have received prior therapy, including an anthracycline, a taxane, a trastuzumab.",
         "raw_biomarkers": "HER2 overexpression",
         "raw_cancer_type": "advanced or metastatic breast cancer",
         "raw_therapeutics": "Tykerb (lapatinib) in combination with capecitabine"
